(PC-31652) refacto after review
